+++ pmd-core/src/main/java/net/sourceforge/pmd/cache/FileAnalysisCache.java
@@ -116,7 +116,7 @@ public class FileAnalysisCache extends AbstractAnalysisCache {
             for (final Map.Entry<String, AnalysisResult> resultEntry : updatedResultsCache.entrySet()) {
                 final List<RuleViolation> violations = resultEntry.getValue().getViolations();
 
-                outputStream.writeUTF(resultEntry.getKey());
+                outputStream.writeUTF(resultEntry.getKey()); // the full filename
                 outputStream.writeLong(resultEntry.getValue().getFileChecksum());
 
                 outputStream.writeInt(violations.size());