+++ pmd-core/src/main/java/net/sourceforge/pmd/renderers/EmacsRenderer.java
@@ -35,7 +35,7 @@ public class EmacsRenderer extends AbstractIncrementingRenderer {
         while (violations.hasNext()) {
             RuleViolation rv = violations.next();
             buf.setLength(0);
-            buf.append(rv.getFilename());
+            buf.append(determineFileName(rv.getFilename()));
             buf.append(':').append(Integer.toString(rv.getBeginLine()));
             buf.append(": ").append(rv.getDescription()).append(EOL);
             writer.write(buf.toString());