+++ docs/pages/release_notes.md
@@ -22,10 +22,23 @@ about the usage and features of the rule designer.
 
 ### Fixed Issues
 
+*   all
+    *   [#1923](https://github.com/pmd/pmd/issues/1923): \[core] Incremental analysis does not work with shortnames
 *   java-codestyle
     *   [#1951](https://github.com/pmd/pmd/issues/1951): \[java] UnnecessaryFullyQualifiedName rule triggered when variable name clashes with package name
 
 ### API Changes
+
+#### Changes to Renderer
+
+*   Each renderer has now a new method {% jdoc core::renderers.Renderer#setUseShortNames(List) %} which
+    is used for implementing the "shortnames" CLI option. The method is automatically called by PMD, if this
+    CLI option is in use. When rendering filenames to the report, the new helper method
+    {% jdoc core::renderers.AbstractRenderer#determineFileName(String) %} should be used. This will change
+    the filename to a short name, if the CLI option "shortnames" is used.
+    
+    Not adjusting custom renderers will make them render always the full file names and not honoring the
+    CLI option "shortnames".
 
 #### Deprecated APIs
 
@@ -34,6 +47,13 @@ about the usage and features of the rule designer.
 *   The methods {% jdoc java::ast.ASTImportDeclaration#getImportedNameNode() %} and
     {% jdoc java::ast.ASTImportDeclaration#getPackage() %} have been deprecated and
     will be removed with PMD 7.0.0.
+*   The method {% jdoc core::RuleContext#setSourceCodeFilename(String) %} has been deprecated
+    and will be removed. The already existing method {% jdoc core::RuleContext#setSourceCodeFile(File)
+    should be used instead. The method {% jdoc core::RuleContext#getSourceCodeFilename() %} still
+    exists and returns just the filename without the full path.
+*   The method {% jdoc core::processor.AbstractPMDProcessor#filenameFrom(DataSource) %} has been
+    deprecated. It was used to determine a "short name" of the file being analyzed, so that the report
+    can use short names. However, this logic has been moved to the renderers.
 
 ### External Contributions